move separate services into one
